Fix build break in C++14 mode: fix incompatible types between declarations and definitions of static data members.
PiperOrigin-RevId: 303714624

+++ riegeli/bytes/cord_backward_writer.cc
@@ -34,7 +34,7 @@ namespace riegeli {
 // namespace scope is required. Since C++17 these definitions are deprecated:
 // http://en.cppreference.com/w/cpp/language/static
 #if __cplusplus < 201703
-constexpr int CordBackwardWriterBase::kShortBufferSize;
+constexpr size_t CordBackwardWriterBase::kShortBufferSize;
 #endif
 
 void CordBackwardWriterBase::Done() {
